backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["marcreniel <marc@notmarc.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
supabase = "^2.7.4"
pydantic = "^2.9.2"
pyjwt = "^2.9.0"
pydantic-settings = "^2.5.2"
celery = "^5.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src:start"
